Shift some compute attributes within request body
The scheduler_hints attribute exists as a top-level key rather than
under the "server" key like most other things, so if we have it set,
shift it out of the resource_key body and into its own top-level key.

The availability_zone attribute exists with two different names,
depending on whether it's in a request or a response. When we're
preparing a request, include it in the request body as
availability_zone, otherwise expect OS-EXT-AZ:availability_zone in the
response body such as for a GET. The same goes for the user_data
attribute, which exists as OS-EXT-SRV-ATTR:user_data in responses.

Change-Id: Ifc261f982f8a13f8ab467350615e08f268a22049
Closes-Bug: 1671073
Closes-Bug: 1671886
(cherry picked from commit feaa7cc7d597b9b04c4c4ebc06f786aebd657559)

Bug-ES #9830
http://192.168.15.2/issues/9830